Rename module `server-env` -> `env-server`.
It now matches the predominant convention and gets ready for the intro of
sibling module `env-client`.
