+++ local-modules/client-bundle/package.json
@@ -15,8 +15,8 @@
     "webpack": "^3.4.1",
 
     "client-tests-loader": "local",
+    "env-server": "local",
     "see-all": "local",
-    "server-env": "local",
     "util-common": "local"
   }
 }
