+++ local-modules/env-server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "server-env",
+  "name": "env-server",
   "version": "1.0.0",
   "main": "main.js",
