+++ server/main.js
@@ -20,10 +20,10 @@ import minimist from 'minimist';
 import { Application } from 'app-setup';
 import { ClientBundle } from 'client-bundle';
 import { DevMode } from 'dev-mode';
+import { Dirs, ProductInfo, ServerEnv } from 'env-server';
 import { Hooks } from 'hooks-server';
 import { Logger } from 'see-all';
 import { FileSink, ServerSink } from 'see-all-server';
-import { Dirs, ProductInfo, ServerEnv } from 'server-env';
 import { ServerTests } from 'testing-server';
 import { PromDelay } from 'util-common';
